Entry/MasterTool refactoring (#36)
