+++ src/app/components/App.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { WorkScreen } from '@/pages/work-screen'
 import { PasswordEntry } from '@/pages/password-entry'
-import { isMasterEntered } from '@/entities/master-password'
+import { isMasterEntered } from '@/entities/master-tool'
 </script>
 
 <template>
