--- src/entities/master-password/index.ts
@@ -1,1 +1,0 @@
-export * from './hooks/master-tool.hook'
