+++ src/entities/master-tool/lib/create-master-tool.ts
