+++ src/entities/master-tool/model/master-tool.ts
