+++ src/entities/master-tool/types.ts
