+++ src/entities/services-config/index.ts
@@ -1,0 +1,1 @@
+export * from './model/service-config'
