+++ src/entities/services-config/lib/create-empty-config.ts
@@ -1,0 +1,6 @@
+import type { ServicesConfig } from '../types'
+
+export const createEmptyConfig = (): ServicesConfig => ({
+  popular: [],
+  custom: [],
+})
