+++ src/entities/services-config/model/service-config.ts
@@ -1,0 +1,14 @@
+import { ref } from 'vue'
+
+import { createEmptyConfig } from '../lib/create-empty-config'
+import type { ServicesConfig } from '../types'
+
+const servicesConfig = ref<ServicesConfig | null>(null)
+
+export const setEmptyServicesConfig = () => {
+  servicesConfig.value = createEmptyConfig()
+}
+
+export const clearServicesConfig = () => {
+  servicesConfig.value = null
+}
