+++ src/features/entry/actions/enter.ts
@@ -1,0 +1,8 @@
+import { setMasterPassword } from '@/entities/master-tool'
+import { setEmptyServicesConfig } from '@/entities/services-config'
+
+export const enter = async (masterPassword: string) => {
+  await setMasterPassword(masterPassword)
+
+  setEmptyServicesConfig()
+}
