+++ src/features/entry/actions/exit.ts
@@ -1,0 +1,7 @@
+import { clearMasterPassword } from '@/entities/master-tool'
+import { clearServicesConfig } from '@/entities/services-config'
+
+export const exit = () => {
+  clearMasterPassword()
+  clearServicesConfig()
+}
