+++ src/features/entry/index.ts
@@ -1,1 +1,3 @@
 export { default as MasterForm } from './components/MasterForm.vue'
+
+export * from './actions/exit'
